exception middleware vs

+++ Core/Base/CoreBase.Dto/IQueryable/Paged/Paging.cs
@@ -1,7 +1,7 @@
 ﻿namespace CoreBase.Dto.IQueryable.Paged;
 
-public class Paging(int page = 1, int size = 10) 
+public class Paging(Int64 page = 1, Int64 size = 10) 
 {
-    public int Page { get; set; } = page > 0 ? page : 1;
-    public int Size { get; set; } = size > 0 ? size : 10;
+    public Int64 Page { get; set; } = page > 0 ? page : 1;
+    public Int64 Size { get; set; } = size > 0 ? size : 10;
 }
